Extract a GitLabPipelineBase class to prepare for the CI-pipeline duration metric.
